extracted SubthemeSelector component
