+++ packages/playground/src/components/index.ts
@@ -3,3 +3,4 @@ export { GeoPosition } from './GeoPosition';
 export { ThemeSelector } from './ThemeSelector';
 export { Selector } from './Selector';
 export { ValidatorSelector } from './ValidatorSelector';
+export { SubthemeSelector } from './SubthemeSelector';
